apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg[binary]==3.2.1
pydantic[email]==2.9.2
python-dotenv==1.0.1
alembic==1.13.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
